chore(coordinator): Move admin HTTP APIs to admin.rs
